Don't display what people have voted until the vote has ended
Currently drek has the ability to see what people have voted when the
vote is open. Removed this and only display it after the voting is closed.

When voting is open, '?' is displayed instead of vote numbers. In order to
see how many people have voted, an indicator to show how many people have
voted has also been added.